requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
transformers
torch
sentencepiece
protobuf